finish the implementation of the v1 cli with signing
