+++ src/account.rs
@@ -1,6 +1,7 @@
-use std::{collections::{HashSet, HashMap}, fmt::Display};
+use std::{collections::HashMap, fmt::Display};
 use serde::{Serialize, Deserialize};
-use crate::quorum::QuorumId;
+use crate::vm_info::VmInfo;
+
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub struct Namespace(String);
@@ -34,6 +35,12 @@ impl TaskId {
     }
 }
 
+impl Display for TaskId {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.task_id())
+    }
+}
+
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum TaskStatus {
     Pending,
@@ -41,30 +48,67 @@ pub enum TaskStatus {
     Failure(String),
 }
 
+impl Display for TaskStatus {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            TaskStatus::Pending => write!(f, "pending"),
+            TaskStatus::Success => write!(f, "success"),
+            TaskStatus::Failure(err_string) => write!(f, "Failure {}", err_string)
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ExposedPort {
+    port: u16,
+    service_description: Option<String>,
+}
+
+impl ExposedPort {
+    pub fn new(port: u16, service_description: Option<String>) -> Self {
+        Self { port, service_description }
+    }
+}
+
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Account {
     address: [u8; 20],
-    namespaces: HashSet<Namespace>,
+    namespaces: HashMap<Namespace, Option<VmInfo>>,
+    exposed_ports: HashMap<Namespace, Vec<ExposedPort>>,
     tasks: HashMap<TaskId, TaskStatus>
 }
 
 impl Account {
     pub fn new(
         address: [u8; 20],
-        namespaces: impl IntoIterator<Item = Namespace>, 
+        namespaces: impl IntoIterator<Item = (Namespace, Option<VmInfo>)>, 
+        exposed_ports: impl IntoIterator<Item = (Namespace, Vec<ExposedPort>)>,
         tasks: impl IntoIterator<Item = (TaskId, TaskStatus)>
     ) -> Self {
         let namespaces = namespaces.into_iter().collect();
         let tasks = tasks.into_iter().collect();
-        Self { address, namespaces, tasks }
+        let exposed_ports = exposed_ports.into_iter().collect();
+        Self { address, namespaces, tasks, exposed_ports }
     }
 
     pub fn address(&self) -> [u8; 20] {
         self.address
     }
 
-    pub fn namespaces(&self) -> HashSet<Namespace> {
+    pub fn namespaces(&self) -> HashMap<Namespace, Option<VmInfo>> {
         self.namespaces.clone()
+    }
+
+    pub fn update_namespace(&mut self, namespace: &Namespace, vminfo: Option<VmInfo>) {
+        self.namespaces.insert(namespace.clone(), vminfo);
+    }
+
+    pub fn update_exposed_ports(&mut self, namespace: &Namespace, exposed_ports: Vec<ExposedPort>) {
+        if let Some(mut entry) = self.exposed_ports.get_mut(namespace) {
+            entry.extend(exposed_ports.into_iter())
+        } else {
+            self.exposed_ports.insert(namespace.clone(), exposed_ports);
+        }
     }
 
     pub fn tasks(&self) -> HashMap<TaskId, TaskStatus> {
@@ -73,10 +117,6 @@ impl Account {
 
     pub fn tasks_mut(&mut self) -> &mut HashMap<TaskId, TaskStatus> {
         &mut self.tasks
-    }
-
-    pub fn get_namespace_quorum(&self, namespace: &Namespace) -> Option<&Namespace> {
-        self.namespaces.get(namespace)
     }
 
     pub fn get_task_status(&self, task_id: &TaskId) -> Option<&TaskStatus> {
