+++ src/client.rs
@@ -1,6 +1,6 @@
 use std::net::SocketAddr;
 use tarpc::{client, context, tokio_serde::formats::Json};
-use allegra::{rpc::VmmClient, vmm::{InstanceCreateParams, InstanceStopParams}, vm_types::VmType};
+use allegra::{rpc::VmmClient, params::{InstanceCreateParams, InstanceStopParams}, vm_types::VmType};
 
 
 #[tokio::main]
@@ -21,27 +21,29 @@ async fn main() -> std::io::Result<()> {
     let mut context = context::current();
     context.deadline = std::time::SystemTime::now() + std::time::Duration::from_secs(60);
 
-//    let res = client.create_vm(
-//        context,
-//        InstanceCreateParams {
-//            name: "testRpcClientServerVm".to_string(),
-//            distro: "ubuntu".to_string(),
-//            version: "22.04".to_string(),
-//            vmtype: VmType::T2Micro, 
-//            sig: "testSignature".to_string()
-//        }
-//    ).await;
+    let res = client.create_vm(
+        context,
+        InstanceCreateParams {
+            name: "testRpcClientServerVm".to_string(),
+            distro: "ubuntu".to_string(),
+            version: "22.04".to_string(),
+            vmtype: VmType::T2Micro, 
+            sig: "testSignature".to_string(),
+            recovery_id: u8::default(),
+        }
+    ).await;
 
-//    println!("{res:#?}");
+    println!("{res:#?}");
 
-//    let mut context = context::current();
-//    context.deadline = std::time::SystemTime::now() + std::time::Duration::from_secs(60);
-//    tokio::time::sleep(tokio::time::Duration::from_secs(240)).await;
+    let mut context = context::current();
+    context.deadline = std::time::SystemTime::now() + std::time::Duration::from_secs(60);
+    tokio::time::sleep(tokio::time::Duration::from_secs(240)).await;
     let res = client.shutdown_vm(
         context,
         InstanceStopParams {
             name: "testRpcClientServerVm".to_string(),
-            sig: "testSignature".to_string()
+            sig: "testSignature".to_string(),
+            recovery_id: u8::default(),
         }
     ).await;
