+++ src/lib.rs
@@ -6,3 +6,6 @@ pub mod account;
 pub mod node;
 pub mod quorum;
 pub mod actors;
+pub mod vm_info;
+pub mod params;
+pub mod helpers;
