+++ start-tikv-server.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+#
+sudo docker run -d --name tikv-server --network host pingcap/tikv:latest \
+  --addr="127.0.0.1:20160" \
+  --advertise-addr="127.0.0.1:20160" \
+  --data-dir="/tikv" \
+  --pd="http://127.0.0.1:2379"
